orderbook api from floatarray to doublearray and orderbook model

+++ app/src/main/java/gts/bitfinex/data/network/BitfinexApi.kt
@@ -28,7 +28,7 @@ interface BitfinexApi {
     fun sendOrderBookRequest(subscribeOrderBook: OrderBookRequest)
 
     @Receive
-    fun observeOrderBook(): Flowable<FloatArray>
+    fun observeOrderBook(): Flowable<DoubleArray>
 
     companion object {
         const val BASE_URI = "wss://api.bitfinex.com/ws/"
