+++ app/src/main/java/gts/bitfinex/domain/model/OrderBookData.kt
@@ -1,12 +1,12 @@
 package gts.bitfinex.domain.model
 
 data class OrderBookData(
-    val price: Float,
+    val price: Double,
     val count: Int,
-    val amount: Float
+    val amount: Double
 )
 
-fun FloatArray.toOrderBookData() = OrderBookData(
+fun DoubleArray.toOrderBookData() = OrderBookData(
     price = this[1],
     count = this[2].toInt(),
     amount = this[3]
